refactor(loader): `HGTLoader` behind `NodeLoader` interface, part 5 (#5418)
This PR builds off of #5404 by refactoring `HGTLoader` behind the sampler and loader interface; the simplicity of this refactor also shows the flexibility of the interface. In doing so, it defines `HGTSampler` that inherits from `BaseSampler`, and uses `HGTSampler` as part of the `HGTLoader(NodeLoader)` construction. With this setup, it should also be trivial to add an HGT link-level loader, but this task is left as a TODO.