Add bipartite `GraphSAGE` example (#5834)
This PR adds an example implementation of a recommendation model
bipartite Graphsage.
The current implementation is a supervised version on MovieLens. The i2i
co-occurrence matrix is applied to construct i2i-graph. When applying
Graphsage, the model follows i-i-i pattern to encode item embedding and
i-i-u for user-side respectively. Since the task to predict u2i rates is
supervised, we apply MSE loss with labels.
For next step, we will apply this model to an unsupervised task on
another dataset with negative sampling and triplet loss.

Co-authored-by: 胡树弦 <hushuxian@B-RR619NXW-1146.local>
Co-authored-by: huxleyhu <hushuxian.hsx@alibaba-inc.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: huxleyhu <shuxian.hu98@gmail.com>
Co-authored-by: Matthias Fey <matthias.fey@tu-dortmund.de>