Fixing Subgraph Methods (#6613)
This pull request addresses a bug in both ```Data.subgraph``` and
```HeteroData.subgraph```, which yield incorrect results when an
unsorted list of integers is used to select a subgraph. In this case,
the attributes of nodes are reordered according to the given list of
integers, but the ids used in ```edge_index``` are *not*. Therefore, the
node features will be distributed incorrectly in the output subgraph.
Note that ```Data.subgraph``` does use the ```utils.subgraph``` method
with the ```relabel_nodes=True``` option to compute the reduced edge
list. However, this is insufficient to ensure a correct output, as this
only shifts node ids downward to ensure consecutive ids. It has no
effect on the actual order of nodes, which causes this problem.

Consider the following example:
```python
import torch
from torch_geometric.data import Data

data = Data(x=torch.tensor([0, 0, 1, 1]), edge_index=torch.tensor([[0, 2], [1, 3]]))
sub = data.subgraph(torch.tensor([0, 2, 1]))
print(sub.x)
print(sub.edge_index)
```
With torch_geometric==2.3.0 (current nightly) we obtain the following
output:
```
x: [0 1 0]
Edge Idx:
 [[0]
 [1]]
```
Note that in the original graph each edge has a source and destination
with the same node label. In the subgraph, the remaining edge has a
source with label 0 and a destination with label 1. This is therefore
not a valid subgraph.

I have addressed this bug by re-mapping the node ids in ```edge_index```
to their correct new index if the input is indeed an unsorted integer
list. ```HeteroData.subgraph``` had the same issue with an analog
solution. I also extended the tests to cover these issues.

Beyond this, I was wondering what the behaviour of ```Data.subgraph```
should be if the argument is a list of integers with duplicate entries.
Should the method raise an Exception, ignore duplicates or create
duplicate vertices? I have not implemented anything in this regard yet,
but I do think this should be specified.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: toensoff <toenshoff@informatik.rwth-aachen.de>
Co-authored-by: Jinu Sunil <jinu.sunil@gmail.com>