chore: move `yamllint` config to pre-commit config (#6776)
Following up with work done in #6431 and #6768, this PR aims to further
improve the overall structure of the repository by moving the yamllint
configuration within the `pre-commit-config.yaml` file.

The only snippet in which the `.yamllint.yml` file was being used was
the pre-commit config file. yamllint [allows to have a custom
configuration without a config
file](https://yamllint.readthedocs.io/en/stable/configuration.html#custom-configuration-without-a-config-file)
by using the `-d` flag. Using this feature, we add a serialized version
of the config in the `args` parameter of the pre-commit config.

---------

Co-authored-by: Matthias Fey <matthias.fey@tu-dortmund.de>