[Transforms] Add `NodePropertySplit` transform (#6894)
New functional for creating node-level data splits, which allows to
induce distributional shifts in graph and conduct experiments with graph
models in more challenging setups. It is implemented via the
`NodePropertySplit` transform and allows to use the class itself or its
static method `mask_nodes_by_property` (the latter is for users who
already knows their structural node property and just need to perform
data splitting).

Test coverage is provided and passed successfully. I decided to edit
`transforms.rst` in `docs/source/modules` to make the static method
`mask_nodes_by_property` visible in documentation, but not sure if this
is the right way.

---------

Co-authored-by: Jintang Li <cnljt@outlook.com>
Co-authored-by: rusty1s <matthias.fey@tu-dortmund.de>