Replace `einsum` with `matmul` in `Performer` (#8035)
The previous implementation uses `torch.einsum` to perform the matrix
multiplications. However, `torch.einsum` is not always as efficient as
`torch.matmul`. The previous implementation needs to perform two
`torch.einsum` calls, one for the inner product and one for the softmax.
The updated implementation uses `torch.matmul` for all of the matrix
multiplications. This is more efficient, especially on GPUs.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: rusty1s <matthias.fey@tu-dortmund.de>